+++ src/test/kotlin/net/pwall/json/JSONSerializerTest.kt
@@ -25,13 +25,13 @@
 
 package net.pwall.json
 
-import java.time.Duration
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
 import kotlin.test.assertSame
 import kotlin.test.assertTrue
 
+import java.time.Duration
 import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalDateTime