+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Changed
+- `JSONDeserializer`: added deserialization of Kotlin objects (serialization already worked)
 
 ## [0.6] - 2019-05-23
 ### Changed