+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Changed
+- `JSONDeserializer`: added deserialization of Kotlin objects (serialization already worked)
+
 ## [0.6] - 2019-05-23
 ### Changed
 - `JSONAuto`: added `JSONConfig` to functions