+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `JSONDeserializer`: added deserialization of Kotlin objects (serialization already worked)
 - `JSONSerializer`: improved serialization of Java objects
 - `JSONDeserializer`: improved serialization of Java objects
+- `JSONSerializer`: added serialization of `java.time.Duration`
 
 ## [0.6] - 2019-05-23
 ### Changed