+++ CHANGELOG.md
@@ -4,10 +4,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 ### Changed
-- `JSONDeserializer`: added deserialization of Kotlin objects (serialization already worked)
+- `JSONDeserializer`: added deserialization of Kotlin `object`s (serialization already worked)
 - `JSONSerializer`: improved serialization of Java objects
-- `JSONDeserializer`: improved serialization of Java objects
+- `JSONDeserializer`: improved deserialization of Java objects
 - `JSONSerializer`: added serialization of `java.time.Duration`
+- `JSONDeserializer`: added deserialization of `java.time.Duration`
+- `JSONConfig`: added `readBufferSize`
 
 ## [0.6] - 2019-05-23
 ### Changed